% Generated by build script
\name{estat_gdp}
\alias{estat_gdp}
\title{Fetch Eurostat GDP data}
\usage{
estat_gdp()
}
\arguments{
\item{countries}{Country codes separated by + (e.g. "DE+FR+IT").}
\item{unit}{Unit of measure: "CLV05_MEUR" (chain-linked volumes),}
\item{na_item}{National accounts item: "B1GQ" (GDP), "P3" (consumption),}
\item{start_period}{Start year}
\item{end_period}{End year}
}
\value{tibble with Eurostat SDMX columns}
\description{Fetch Eurostat GDP data}
