% Generated by build script
\name{estat_inflation}
\alias{estat_inflation}
\title{Fetch Eurostat inflation data (HICP)}
\usage{
estat_inflation()
}
\arguments{
\item{countries}{Country codes (e.g. "DE+FR")}
\item{start_period}{Start year}
\item{end_period}{End year}
}
\value{tibble}
\description{Fetch Eurostat inflation data (HICP)}
