% Generated by build script
\name{estat_unemployment}
\alias{estat_unemployment}
\title{Fetch Eurostat unemployment data}
\usage{
estat_unemployment()
}
\arguments{
\item{countries}{Country codes (e.g. "DE+FR")}
\item{start_period}{Start period (e.g. "2020-01")}
\item{end_period}{End period}
}
\value{tibble}
\description{Fetch Eurostat unemployment data}
