% Generated by build script
\name{pvgis_monthly}
\alias{pvgis_monthly}
\title{Get monthly PV energy output estimates}
\usage{
pvgis_monthly()
}
\arguments{
\item{lat}{Latitude}
\item{lon}{Longitude}
\item{peakpower}{Peak power of PV system in kW (default 1)}
}
\value{tibble of monthly estimates}
\description{Get monthly PV energy output estimates}
