% Generated by build script
\name{pvgis_optimal}
\alias{pvgis_optimal}
\title{Get optimal PV system angle and yearly output}
\usage{
pvgis_optimal()
}
\arguments{
\item{lat}{Latitude}
\item{lon}{Longitude}
\item{peakpower}{Peak power in kW (default 1)}
\item{loss}{System loss in percent (default 14)}
}
\value{tibble: one row with optimal_angle, optimal_azimuth, yearly_energy_kwh,}
\description{Get optimal PV system angle and yearly output}
