% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estat.R
\name{estat_inflation}
\alias{estat_inflation}
\title{Fetch Eurostat inflation data (HICP)}
\usage{
estat_inflation(countries = "", start_period = NULL, end_period = NULL)
}
\arguments{
\item{countries}{Country codes (e.g. "DE+FR")}

\item{start_period}{Start year}

\item{end_period}{End year}
}
\value{
tibble
}
\description{
Annual harmonised index of consumer prices from prc_hicp_aind.
}
