% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estat.R
\name{estat_unemployment}
\alias{estat_unemployment}
\title{Fetch Eurostat unemployment data}
\usage{
estat_unemployment(countries = "", start_period = NULL, end_period = NULL)
}
\arguments{
\item{countries}{Country codes (e.g. "DE+FR")}

\item{start_period}{Start period (e.g. "2020-01")}

\item{end_period}{End period}
}
\value{
tibble
}
\description{
Monthly unemployment rate from une_rt_m.
}
